requirements.txt
streamlit==1.40.2
bcrypt==4.2.0
python-dotenv==1.0.1
openpyxl==3.1.5
streamlit-cookies-manager==0.2.0
